chore(deslop): break two real circular dependencies
madge --circular on core/ui/agent/app-core surfaced 6 cycles. Four are
deliberate/false-positive (type-only imports or intentional dynamic-import
seams): core entities.ts<->roles.ts (type-only + lazy await import),
ui DynamicViewLoader dynamic import("../../index.ts") manifest broker,
plugin-local-inference type-only edge. Two were real barrel/wrong-direction
edges and are broken here:

- app-core api/auth.ts <-> api/auth/auth-context.ts: auth-context imported
  getProvidedApiToken from the auth.ts barrel while auth.ts re-exported
  auth-context's session helpers. Moved the pure header-parsing helpers
  (extractHeaderValue, getProvidedApiToken) down into the leaf api/auth/tokens.ts;
  auth.ts re-exports them (public API unchanged) and auth-context imports from
  ./tokens.js. Cycle gone.

- ui cloud/handoff/resume-pending-handoff.ts pulled loadPersistedActiveServer
  from the ../../state barrel, closing a state/index -> AppContext ->
  useStartupCoordinator -> startup-phase-poll -> resume-pending-handoff loop.
  Import directly from ../../state/persistence instead. Cycle gone.

Verified: madge re-run shows both cycles removed; app-core tsgo --noEmit clean
on touched files (only pre-existing missing-generated-file errors remain);
biome check clean.

The account-pool<->coding-account-bridge cycle is a genuine bidirectional
runtime dependency needing a design inversion; left as-is and reported.
