+++ packages/app-core/src/api/auth/tokens.ts
@@ -5,6 +5,7 @@
  * compare and folds the true length check into the returned boolean.
  */
 import crypto from "node:crypto";
+import type http from "node:http";
 
 /** Timing-safe token comparison (constant-time regardless of input length). */
 export function tokenMatches(expected: string, provided: string): boolean {
@@ -17,4 +18,44 @@ export function tokenMatches(expected: string, provided: string): boolean {
   b.copy(bPadded);
   const contentMatch = crypto.timingSafeEqual(aPadded, bPadded);
   return a.length === b.length && contentMatch;
+}
+
+/**
+ * Normalise a potentially multi-valued HTTP header into a single string.
+ * Returns `null` when the header is absent or empty.
+ */
+export function extractHeaderValue(
+  value: string | string[] | undefined,
+): string | null {
+  if (typeof value === "string") return value;
+  return Array.isArray(value) && typeof value[0] === "string" ? value[0] : null;
+}
+
+/**
+ * Extract the API token from an incoming request.
+ *
+ * Checks (in order):
+ *   1. `Authorization: Bearer <token>`
+ *   2. `x-eliza-token`
+ *   3. `x-elizaos-token`
+ *   4. `x-api-key` / `x-api-token`
+ */
+export function getProvidedApiToken(
+  req: Pick<http.IncomingMessage, "headers">,
+): string | null {
+  const authHeader = extractHeaderValue(req.headers.authorization)
+    ?.slice(0, 1024)
+    ?.trim();
+  if (authHeader) {
+    const match = /^Bearer\s{1,8}(.+)$/i.exec(authHeader);
+    if (match?.[1]) return match[1].trim();
+  }
+
+  const headerToken =
+    extractHeaderValue(req.headers["x-eliza-token"]) ??
+    extractHeaderValue(req.headers["x-elizaos-token"]) ??
+    extractHeaderValue(req.headers["x-api-key"]) ??
+    extractHeaderValue(req.headers["x-api-token"]);
+
+  return headerToken?.trim() || null;
 }
